feat(main): Add virtio-block device support, adjust ramdisk interface content, and unify the external API interface of boot media
- Add virtio-block device support, implemented by encapsulating related interfaces of axfs
- Modify ramdisk-cpio related interfaces and open them to the public
- Shield the differences between different boot media and unify the medium API interface
