+++ src/main.rs
@@ -6,7 +6,7 @@ extern crate axlog;
 
 mod panic;
 mod log;
-mod media;
+mod medium;
 
 #[cfg_attr(not(test), unsafe(no_mangle))]
 pub extern "C" fn rust_main(_cpu_id: usize, _dtb: usize) -> ! {
@@ -42,7 +42,7 @@ pub extern "C" fn rust_main(_cpu_id: usize, _dtb: usize) -> ! {
         #[cfg(feature = "fs")]
         // 目前使用ramdisk的cpio格式时，驱动还不完善，用不了，需要注释掉
         // 如果使用virtio-blk驱动，则可以正常使用
-        //axfs::init_filesystems(all_devices.block);
+        axfs::init_filesystems(all_devices.block);
 
         #[cfg(feature = "net")]
         axnet::init_network(all_devices.net);
@@ -52,9 +52,8 @@ pub extern "C" fn rust_main(_cpu_id: usize, _dtb: usize) -> ! {
     }
     ctor_bare::call_ctors();
 
-    info!("will test cpio.");
-
-    crate::media::parse_cpio_ramdisk();
+    info!("current root dir: {}", crate::medium::current_dir().unwrap());
+    info!("read test file context: {}", crate::medium::read_to_string("/test/arceboot.txt").unwrap());
 
     info!("will shut down.");
